backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloud-sql-python-connector==0.6.0
cryptography==36.0.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
mysql-connector-python==8.0.28
numpy
pandas
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2
